added barcode scanner example
