+++ examples/basic/android/settings.gradle
@@ -1,4 +1,6 @@
 rootProject.name = 'RNCameraExample'
+include ':react-native-gesture-handler'
+project(':react-native-gesture-handler').projectDir = new File(rootProject.projectDir, '../node_modules/react-native-gesture-handler/android')
 include ':react-native-camera'
 project(':react-native-camera').projectDir = new File(rootProject.projectDir, '../../../android')
